requirements.txt
python-dotenv
deepgram-sdk
pyaudio
pyperclip
pygame